+++ deps.sh
@@ -1,6 +1,0 @@
-set -e
-
-# install non-code dependencies used for the project
-
-cargo install simple-http-server
-cargo install cargo-bundle