+++ serve_wasm.sh
@@ -1,5 +1,0 @@
-set -e
-
-./build_wasm.sh
-cd web
-simple-http-server --index -p ${PORT:=5577} .